@@ -2,11 +2,12 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="..\Test\TestLogger.cs" Link="TestLogger.cs" />
+    <Compile Include="..\Test\Generated\TestService.Generated.cs" Link="TestService.Generated.cs" />
   </ItemGroup>
 
   <ItemGroup>
