@@ -1,16 +1,14 @@
-namespace RpcNet.Internal
+// Copyright by Artur Wolf
+
+namespace RpcNet.Internal;
+
+using System.Net.Sockets;
+
+// Public for tests
+public readonly struct NetworkWriteResult
 {
-    using System.Net.Sockets;
+    public NetworkWriteResult(SocketError socketError) => SocketError = socketError;
 
-    // Public for tests
-    public readonly struct NetworkWriteResult
-    {
-        public NetworkWriteResult(SocketError socketError)
-        {
-            this.SocketError = socketError;
-        }
-
-        public SocketError SocketError { get; }
-        public bool HasError => this.SocketError != SocketError.Success;
-    }
+    public SocketError SocketError { get; }
+    public bool HasError => SocketError != SocketError.Success;
 }
