@@ -1,150 +1,156 @@
-namespace RpcNet.Internal
+// Copyright by Artur Wolf
+
+namespace RpcNet.Internal;
+
+using System.Net;
+
+internal class RpcCall
 {
-    using System;
-    using System.Net;
+    private readonly ILogger _logger;
+    private readonly INetworkReader _networkReader;
+    private readonly INetworkWriter _networkWriter;
+    private readonly Action _reestablishConnection;
+    private readonly IPEndPoint _remoteIpEndPoint;
+    private readonly RpcMessage _rpcMessage;
+    private readonly IXdrReader _xdrReader;
+    private readonly IXdrWriter _xdrWriter;
 
-    internal class RpcCall
+    private uint _nextXid = (uint)new Random().Next();
+
+    public RpcCall(
+        int program,
+        IPEndPoint remoteIpEndPoint,
+        INetworkReader networkReader,
+        INetworkWriter networkWriter,
+        Action reestablishConnection,
+        ILogger logger = default)
     {
-        private readonly ILogger logger;
-        private readonly INetworkReader networkReader;
-        private readonly INetworkWriter networkWriter;
-        private readonly Action reestablishConnection;
-        private readonly IPEndPoint remoteIpEndPoint;
-        private readonly RpcMessage rpcMessage;
-        private readonly IXdrReader xdrReader;
-        private readonly IXdrWriter xdrWriter;
-
-        private uint nextXid = (uint)new Random().Next();
-
-        public RpcCall(
-            int program,
-            IPEndPoint remoteIpEndPoint,
-            INetworkReader networkReader,
-            INetworkWriter networkWriter,
-            Action reestablishConnection,
-            ILogger logger = default)
+        _remoteIpEndPoint = remoteIpEndPoint;
+        _networkReader = networkReader;
+        _networkWriter = networkWriter;
+        _xdrReader = new XdrReader(networkReader);
+        _xdrWriter = new XdrWriter(networkWriter);
+        _rpcMessage = new RpcMessage
         {
-            this.remoteIpEndPoint = remoteIpEndPoint;
-            this.networkReader = networkReader;
-            this.networkWriter = networkWriter;
-            this.xdrReader = new XdrReader(networkReader);
-            this.xdrWriter = new XdrWriter(networkWriter);
-            this.rpcMessage = new RpcMessage
+            Body =
             {
-                Body = new Body
+                MessageType = MessageType.Call,
+                CallBody =
                 {
-                    MessageType = MessageType.Call,
-                    CallBody = new CallBody
+                    RpcVersion = 2,
+                    Program = (uint)program,
+                    Credential =
                     {
-                        RpcVersion = 2,
-                        Program = (uint)program,
-                        Credential = new OpaqueAuthentication
-                            { AuthenticationFlavor = AuthenticationFlavor.None, Body = new byte[0] },
-                        Verifier = new OpaqueAuthentication
-                            { AuthenticationFlavor = AuthenticationFlavor.None, Body = new byte[0] }
+                        AuthenticationFlavor = AuthenticationFlavor.None,
+                        Body = Array.Empty<byte>()
+                    },
+                    Verifier =
+                    {
+                        AuthenticationFlavor = AuthenticationFlavor.None,
+                        Body = Array.Empty<byte>()
                     }
                 }
-            };
-            this.logger = logger;
-            this.reestablishConnection = reestablishConnection;
-        }
+            }
+        };
+        _logger = logger;
+        _reestablishConnection = reestablishConnection;
+    }
 
-        public void SendCall(int procedure, int version, IXdrWritable argument, IXdrReadable result)
+    public void SendCall(int procedure, int version, IXdrDataType argument, IXdrDataType result)
+    {
+        for (int i = 0; i < 2; i++)
         {
-            for (int i = 0; i < 2; i++)
+            if (!SendMessage(procedure, version, argument, out string errorMessage))
             {
-                if (!this.SendMessage(procedure, version, argument, out string errorMessage))
+                if (i == 0)
                 {
-                    if (i == 0)
-                    {
-                        this.logger?.Error(errorMessage + " Retrying...");
-                        this.reestablishConnection?.Invoke();
-                        continue;
-                    }
-
-                    throw new RpcException(errorMessage);
+                    _logger?.Error(errorMessage + " Retrying...");
+                    _reestablishConnection?.Invoke();
+                    continue;
                 }
 
-                if (!this.ReceiveReply(result, out errorMessage))
-                {
-                    if (i == 0)
-                    {
-                        this.logger?.Error(errorMessage + " Retrying...");
-                        this.reestablishConnection?.Invoke();
-                        continue;
-                    }
+                throw new RpcException(errorMessage);
+            }
 
-                    throw new RpcException(errorMessage);
+            if (!ReceiveReply(result, out errorMessage))
+            {
+                if (i == 0)
+                {
+                    _logger?.Error(errorMessage + " Retrying...");
+                    _reestablishConnection?.Invoke();
+                    continue;
                 }
 
-                break;
+                throw new RpcException(errorMessage);
             }
-        }
 
-        private bool SendMessage(int procedure, int version, IXdrWritable argument, out string errorMessage)
+            break;
+        }
+    }
+
+    private bool SendMessage(int procedure, int version, IXdrDataType argument, out string errorMessage)
+    {
+        _networkWriter.BeginWriting();
+
+        _rpcMessage.Xid = _nextXid++;
+        _rpcMessage.Body.CallBody.Procedure = (uint)procedure;
+        _rpcMessage.Body.CallBody.Version = (uint)version;
+        _rpcMessage.WriteTo(_xdrWriter);
+        argument.WriteTo(_xdrWriter);
+
+        NetworkWriteResult writeResult = _networkWriter.EndWriting(_remoteIpEndPoint);
+        if (writeResult.HasError)
         {
-            this.networkWriter.BeginWriting();
-
-            this.rpcMessage.Xid = this.nextXid++;
-            this.rpcMessage.Body.CallBody.Procedure = (uint)procedure;
-            this.rpcMessage.Body.CallBody.Version = (uint)version;
-            this.rpcMessage.WriteTo(this.xdrWriter);
-            argument.WriteTo(this.xdrWriter);
-
-            NetworkWriteResult writeResult = this.networkWriter.EndWriting(this.remoteIpEndPoint);
-            if (writeResult.HasError)
-            {
-                errorMessage =
-                    $"Could not send message to {this.remoteIpEndPoint}. Socket error: {writeResult.SocketError}.";
-                return false;
-            }
-
-            NetworkReadResult readResult = this.networkReader.BeginReading();
-            if (readResult.HasError)
-            {
-                errorMessage =
-                    $"Could not receive reply from {this.remoteIpEndPoint}. Socket error: {readResult.SocketError}.";
-                return false;
-            }
-
-            errorMessage = null;
-            return true;
+            errorMessage =
+                $"Could not send message to {_remoteIpEndPoint}. Socket error: {writeResult.SocketError}.";
+            return false;
         }
 
-        private bool ReceiveReply(IXdrReadable result, out string errorMessage)
+        NetworkReadResult readResult = _networkReader.BeginReading();
+        if (readResult.HasError)
         {
-            var reply = new RpcMessage();
-            reply.ReadFrom(this.xdrReader);
-
-            if (reply.Xid != this.rpcMessage.Xid)
-            {
-                errorMessage = $"Wrong XID. Expected {this.rpcMessage.Xid}, but was {reply.Xid}.";
-                return false;
-            }
-
-            if (reply.Body.MessageType != MessageType.Reply)
-            {
-                errorMessage = $"Wrong message type. Expected {MessageType.Reply}, but was {reply.Body.MessageType}.";
-                return false;
-            }
-
-            if (reply.Body.ReplyBody.ReplyStatus != ReplyStatus.Accepted)
-            {
-                errorMessage = "Call was denied.";
-                return false;
-            }
-
-            if (reply.Body.ReplyBody.AcceptedReply.ReplyData.AcceptStatus != AcceptStatus.Success)
-            {
-                errorMessage = $"Call was unsuccessful: {reply.Body.ReplyBody.AcceptedReply.ReplyData.AcceptStatus}.";
-                return false;
-            }
-
-            result.ReadFrom(this.xdrReader);
-            this.networkReader.EndReading();
-
-            errorMessage = null;
-            return true;
+            errorMessage =
+                $"Could not receive reply from {_remoteIpEndPoint}. Socket error: {readResult.SocketError}.";
+            return false;
         }
+
+        errorMessage = null;
+        return true;
+    }
+
+    private bool ReceiveReply(IXdrDataType result, out string errorMessage)
+    {
+        var reply = new RpcMessage();
+        reply.ReadFrom(_xdrReader);
+
+        if (reply.Xid != _rpcMessage.Xid)
+        {
+            errorMessage = $"Wrong XID. Expected {_rpcMessage.Xid}, but was {reply.Xid}.";
+            return false;
+        }
+
+        if (reply.Body.MessageType != MessageType.Reply)
+        {
+            errorMessage = $"Wrong message type. Expected {MessageType.Reply}, but was {reply.Body.MessageType}.";
+            return false;
+        }
+
+        if (reply.Body.ReplyBody.ReplyStatus != ReplyStatus.Accepted)
+        {
+            errorMessage = "Call was denied.";
+            return false;
+        }
+
+        if (reply.Body.ReplyBody.AcceptedReply.ReplyData.AcceptStatus != AcceptStatus.Success)
+        {
+            errorMessage = $"Call was unsuccessful: {reply.Body.ReplyBody.AcceptedReply.ReplyData.AcceptStatus}.";
+            return false;
+        }
+
+        result.ReadFrom(_xdrReader);
+        _networkReader.EndReading();
+
+        errorMessage = null;
+        return true;
     }
 }
