@@ -1,89 +1,99 @@
-namespace RpcNet.Internal
+// Copyright by Artur Wolf
+
+namespace RpcNet.Internal;
+
+using System.Net;
+using System.Net.Sockets;
+using PortMapper;
+
+// Public for tests
+public class RpcTcpClient : INetworkClient
 {
-    using System;
-    using System.Net;
-    using System.Net.Sockets;
-    using RpcNet.PortMapper;
+    private readonly RpcCall _call;
+    private readonly IPEndPoint _remoteIpEndPoint;
+    private readonly TcpReader _tcpReader;
+    private readonly TcpWriter _tcpWriter;
 
-    // Public for tests
-    public class RpcTcpClient : INetworkClient
+    private Socket _client;
+
+    public RpcTcpClient(
+        IPAddress ipAddress,
+        int port,
+        int program,
+        int version,
+        ClientSettings clientSettings = default)
     {
-        private readonly RpcCall call;
-        private readonly IPEndPoint remoteIpEndPoint;
-        private readonly TcpReader tcpReader;
-        private readonly TcpWriter tcpWriter;
+        clientSettings ??= new ClientSettings();
 
-        private Socket client;
-
-        public RpcTcpClient(
-            IPAddress ipAddress,
-            int program,
-            int version,
-            ClientSettings clientSettings = default)
+        if (port == 0)
         {
-            int port = clientSettings?.Port ?? 0;
-            if (port == 0)
+            if (program == PortMapperConstants.PortMapperProgram)
             {
-                var portMapperClientSettings = new PortMapperClientSettings
-                {
-                    Port = clientSettings?.PortMapperPort ?? PortMapperConstants.PortMapperPort
-                };
-                port = PortMapperUtilities.GetPort(Protocol.Tcp, ipAddress, program, version, portMapperClientSettings);
+                port = PortMapperConstants.PortMapperPort;
             }
-
-            this.remoteIpEndPoint = new IPEndPoint(ipAddress, port);
-            this.client = new Socket(SocketType.Stream, ProtocolType.Tcp);
-            this.ReceiveTimeout = clientSettings?.ReceiveTimeout ?? Utilities.DefaultClientReceiveTimeout;
-            this.SendTimeout = clientSettings?.SendTimeout ?? Utilities.DefaultClientSendTimeout;
-            this.EstablishConnection();
-            this.tcpReader = new TcpReader(this.client, clientSettings?.Logger);
-            this.tcpWriter = new TcpWriter(this.client, clientSettings?.Logger);
-            this.call = new RpcCall(
-                program,
-                this.remoteIpEndPoint,
-                this.tcpReader,
-                this.tcpWriter,
-                this.ReestablishConnection,
-                clientSettings?.Logger);
-        }
-
-        public TimeSpan ReceiveTimeout
-        {
-            get => Utilities.GetReceiveTimeout(this.client);
-            set => Utilities.SetReceiveTimeout(this.client, value);
-        }
-
-        public TimeSpan SendTimeout
-        {
-            get => Utilities.GetSendTimeout(this.client);
-            set => Utilities.SetSendTimeout(this.client, value);
-        }
-
-        public void Call(int procedure, int version, IXdrWritable argument, IXdrReadable result) =>
-            this.call.SendCall(procedure, version, argument, result);
-
-        public void Dispose() => this.client.Dispose();
-
-        private void EstablishConnection()
-        {
-            try
+            else
             {
-                this.client.Connect(this.remoteIpEndPoint);
-            }
-            catch (SocketException e)
-            {
-                throw new RpcException(
-                    $"Could not connect to {this.remoteIpEndPoint}. Socket error: {e.SocketErrorCode}.");
+                port = PortMapperUtilities.GetPort(
+                    ProtocolKind.Tcp,
+                    ipAddress,
+                    clientSettings.PortMapperPort,
+                    program,
+                    version,
+                    clientSettings);
             }
         }
 
-        private void ReestablishConnection()
+        _remoteIpEndPoint = new IPEndPoint(ipAddress, port);
+        _client = new Socket(SocketType.Stream, ProtocolType.Tcp);
+        ReceiveTimeout = clientSettings.ReceiveTimeout;
+        SendTimeout = clientSettings.SendTimeout;
+        EstablishConnection();
+        _tcpReader = new TcpReader(_client, clientSettings.Logger);
+        _tcpWriter = new TcpWriter(_client, clientSettings.Logger);
+        _call = new RpcCall(
+            program,
+            _remoteIpEndPoint,
+            _tcpReader,
+            _tcpWriter,
+            ReestablishConnection,
+            clientSettings.Logger);
+    }
+
+    public TimeSpan ReceiveTimeout
+    {
+        get => Utilities.GetReceiveTimeout(_client);
+        set => Utilities.SetReceiveTimeout(_client, value);
+    }
+
+    public TimeSpan SendTimeout
+    {
+        get => Utilities.GetSendTimeout(_client);
+        set => Utilities.SetSendTimeout(_client, value);
+    }
+
+    public void Call(int procedure, int version, IXdrDataType argument, IXdrDataType result) =>
+        _call.SendCall(procedure, version, argument, result);
+
+    public void Dispose() => _client.Dispose();
+
+    private void EstablishConnection()
+    {
+        try
         {
-            this.client.Close();
-            this.client = new Socket(SocketType.Stream, ProtocolType.Tcp);
-            this.EstablishConnection();
-            this.tcpReader.Reset(this.client);
-            this.tcpWriter.Reset(this.client);
+            _client.Connect(_remoteIpEndPoint);
         }
+        catch (SocketException e)
+        {
+            throw new RpcException($"Could not connect to {_remoteIpEndPoint}. Socket error: {e.SocketErrorCode}.");
+        }
+    }
+
+    private void ReestablishConnection()
+    {
+        _client.Close();
+        _client = new Socket(SocketType.Stream, ProtocolType.Tcp);
+        EstablishConnection();
+        _tcpReader.Reset(_client);
+        _tcpWriter.Reset(_client);
     }
 }
