@@ -1,126 +1,145 @@
-namespace RpcNet.Internal
+// Copyright by Artur Wolf
+
+namespace RpcNet.Internal;
+
+using System.Net;
+using System.Net.Sockets;
+using PortMapper;
+
+// Public for tests
+public class RpcUdpServer : IDisposable
 {
-    using System;
-    using System.Net;
-    using System.Net.Sockets;
-    using System.Threading;
-    using RpcNet.PortMapper;
+    private readonly ILogger _logger;
+    private readonly int _port;
+    private readonly UdpReader _reader;
+    private readonly ReceivedRpcCall _receivedCall;
+    private readonly Socket _server;
+    private readonly UdpWriter _writer;
 
-    // Public for tests
-    public class RpcUdpServer : IDisposable
+    private bool _isDisposed;
+    private Thread _receivingThread;
+    private volatile bool _stopReceiving;
+
+    public RpcUdpServer(
+        IPAddress ipAddress,
+        int port,
+        int program,
+        int[] versions,
+        Action<ReceivedRpcCall> receivedCallDispatcher,
+        ServerSettings serverSettings = default)
     {
-        private readonly ILogger logger;
-        private readonly int port;
-        private readonly UdpReader reader;
-        private readonly ReceivedRpcCall receivedCall;
-        private readonly Socket server;
-        private readonly UdpWriter writer;
+        serverSettings ??= new ServerSettings();
 
-        private bool isDisposed;
-        private Thread receivingThread;
-        private volatile bool stopReceiving;
+        _port = port;
+        _server = new Socket(AddressFamily.InterNetwork, SocketType.Dgram, ProtocolType.Udp);
+        _server.Bind(new IPEndPoint(ipAddress, _port));
 
-        public RpcUdpServer(
-            IPAddress ipAddress,
-            int program,
-            int[] versions,
-            Action<ReceivedRpcCall> receivedCallDispatcher,
-            ServerSettings serverSettings = default)
+        _reader = new UdpReader(_server);
+        _writer = new UdpWriter(_server);
+
+        _receivedCall = new ReceivedRpcCall(program, versions, _reader, _writer, receivedCallDispatcher);
+
+        _logger = serverSettings.Logger;
+
+        if (_port == 0)
         {
-            this.port = serverSettings?.Port ?? 0;
-            this.server = new Socket(AddressFamily.InterNetwork, SocketType.Dgram, ProtocolType.Udp);
-            this.server.Bind(new IPEndPoint(ipAddress, this.port));
-
-            this.reader = new UdpReader(this.server);
-            this.writer = new UdpWriter(this.server);
-
-            this.receivedCall = new ReceivedRpcCall(
-                program,
-                versions,
-                this.reader,
-                this.writer,
-                receivedCallDispatcher);
-
-            this.logger = serverSettings?.Logger;
-
-            if (this.port == 0)
+            if (_server.LocalEndPoint is not IPEndPoint localEndPoint)
             {
-                this.port = ((IPEndPoint)this.server.LocalEndPoint).Port;
-                var clientSettings = new PortMapperClientSettings
-                    { Port = serverSettings?.PortMapperPort ?? PortMapperConstants.PortMapperPort };
+                throw new InvalidOperationException("Could not find local endpoint for server socket.");
+            }
+
+            _port = localEndPoint.Port;
+
+            if (program != PortMapperConstants.PortMapperProgram)
+            {
+                var clientSettings = new ClientSettings
+                {
+                    Logger = serverSettings.Logger,
+                    ReceiveTimeout = serverSettings.ReceiveTimeout,
+                    SendTimeout = serverSettings.SendTimeout
+                };
                 foreach (int version in versions)
                 {
-                    PortMapperUtilities.UnsetAndSetPort(Protocol.Udp, this.port, program, version, clientSettings);
+                    PortMapperUtilities.UnsetAndSetPort(
+                        ProtocolKind.Udp,
+                        serverSettings.PortMapperPort,
+                        _port,
+                        program,
+                        version,
+                        clientSettings);
                 }
             }
-
-            this.logger?.Info(
-                $"{Utilities.ConvertToString(Protocol.Udp)} Server listening on {this.server.LocalEndPoint}...");
         }
 
-        public void Start()
+        _logger?.Info($"{Utilities.ConvertToString(Protocol.Udp)} Server listening on {_server.LocalEndPoint}...");
+    }
+
+    public int Start()
+    {
+        if (_isDisposed)
         {
-            if (this.isDisposed)
-            {
-                throw new ObjectDisposedException(nameof(RpcUdpServer));
-            }
-
-            if (this.receivingThread != null)
-            {
-                return;
-            }
-
-            this.receivingThread = new Thread(this.HandlingUdpCalls)
-                { IsBackground = true, Name = $"RpcNet UDP {this.port}" };
-            this.receivingThread.Start();
+            throw new ObjectDisposedException(nameof(RpcUdpServer));
         }
 
-        public void Dispose()
+        if (_receivingThread != null)
         {
-            this.stopReceiving = true;
+            return _port;
+        }
+
+        _receivingThread = new Thread(HandlingUdpCalls)
+        {
+            IsBackground = true,
+            Name = $"RpcNet UDP Server Thread for Port {_port}"
+        };
+        _receivingThread.Start();
+        return _port;
+    }
+
+    public void Dispose()
+    {
+        _stopReceiving = true;
+        try
+        {
+            // Necessary for Linux. Dispose doesn't abort synchronous calls
+            _server.Shutdown(SocketShutdown.Both);
+        }
+        catch
+        {
+            // Ignored
+        }
+
+        _server.Dispose();
+        Interlocked.Exchange(ref _receivingThread, null)?.Join();
+        _isDisposed = true;
+    }
+
+    private void HandlingUdpCalls()
+    {
+        while (!_stopReceiving)
+        {
             try
             {
-                // Necessary for Linux. Dispose doesn't abort synchronous calls
-                this.server.Shutdown(SocketShutdown.Both);
-            }
-            catch
-            {
-                // Ignored
-            }
-
-            this.server.Dispose();
-            Interlocked.Exchange(ref this.receivingThread, null)?.Join();
-            this.isDisposed = true;
-        }
-
-        private void HandlingUdpCalls()
-        {
-            while (!this.stopReceiving)
-            {
-                try
+                NetworkReadResult result = _reader.BeginReading();
+                if (result.HasError)
                 {
-                    NetworkReadResult result = this.reader.BeginReading();
-                    if (result.HasError)
-                    {
-                        this.logger?.Trace($"Could not read UDP data. Socket error: {result.SocketError}.");
-                        continue;
-                    }
-
-                    if (result.IsDisconnected)
-                    {
-                        // Should only happen on dispose
-                        continue;
-                    }
-
-                    this.writer.BeginWriting();
-                    this.receivedCall.HandleCall(new Caller(result.RemoteIpEndPoint, Protocol.Udp));
-                    this.reader.EndReading();
-                    this.writer.EndWriting(result.RemoteIpEndPoint);
+                    _logger?.Trace($"Could not read UDP data. Socket error: {result.SocketError}.");
+                    continue;
                 }
-                catch (Exception e)
+
+                if (result.IsDisconnected)
                 {
-                    this.logger?.Error($"The following error occurred while processing UDP call: {e}");
+                    // Should only happen on dispose
+                    continue;
                 }
+
+                _writer.BeginWriting();
+                _receivedCall.HandleCall(new Caller(result.RemoteIpEndPoint, Protocol.Udp));
+                _reader.EndReading();
+                _writer.EndWriting(result.RemoteIpEndPoint);
+            }
+            catch (Exception e)
+            {
+                _logger?.Error($"The following error occurred while processing UDP call: {e}");
             }
         }
     }
