@@ -1,83 +1,83 @@
-namespace RpcNet.Internal
+// Copyright by Artur Wolf
+
+namespace RpcNet.Internal;
+
+using System.Net;
+using System.Net.Sockets;
+using System.Runtime.InteropServices;
+
+// Public for tests
+public class UdpReader : INetworkReader
 {
-    using System;
-    using System.Net;
-    using System.Net.Sockets;
-    using System.Runtime.InteropServices;
+    private readonly byte[] _buffer;
+    private readonly Socket _udpClient;
 
-    // Public for tests
-    public class UdpReader : INetworkReader
+    private int _readIndex;
+    private EndPoint _remoteEndPoint = new IPEndPoint(IPAddress.Loopback, 0);
+    private int _totalLength;
+
+    public UdpReader(Socket udpClient) : this(udpClient, 65536)
     {
-        private readonly byte[] buffer;
-        private readonly Socket udpClient;
+    }
 
-        private int readIndex;
-        private EndPoint remoteEndPoint = new IPEndPoint(IPAddress.Loopback, 0);
-        private int totalLength;
-
-        public UdpReader(Socket udpClient) : this(udpClient, 65536)
+    public UdpReader(Socket udpClient, int bufferSize)
+    {
+        if ((bufferSize < sizeof(int)) || ((bufferSize % 4) != 0))
         {
+            throw new ArgumentOutOfRangeException(nameof(bufferSize));
         }
 
-        public UdpReader(Socket udpClient, int bufferSize)
+        _udpClient = udpClient;
+
+        // See
+        // https://stackoverflow.com/questions/7201862/an-existing-connection-was-forcibly-closed-by-the-remote-host
+        if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
         {
-            if ((bufferSize < sizeof(int)) || ((bufferSize % 4) != 0))
-            {
-                throw new ArgumentOutOfRangeException(nameof(bufferSize));
-            }
-
-            this.udpClient = udpClient;
-
-            // See
-            // https://stackoverflow.com/questions/7201862/an-existing-connection-was-forcibly-closed-by-the-remote-host
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-            {
-                const uint IOC_IN = 0x80000000;
-                const uint IOC_VENDOR = 0x18000000;
-                uint SIO_UDP_CONNRESET = IOC_IN | IOC_VENDOR | 12;
-                this.udpClient.IOControl((int)SIO_UDP_CONNRESET, new[] { Convert.ToByte(false) }, null);
-            }
-
-            this.buffer = new byte[bufferSize];
+            const uint IocIn = 0x80000000;
+            const uint IocVendor = 0x18000000;
+            const uint SioUdpConnectionReset = IocIn | IocVendor | 12;
+            _udpClient.IOControl(unchecked((int)SioUdpConnectionReset), new[] { Convert.ToByte(false) }, null);
         }
 
-        public NetworkReadResult BeginReading()
-        {
-            this.readIndex = 0;
-            try
-            {
-                this.totalLength = this.udpClient.ReceiveFrom(this.buffer, SocketFlags.None, ref this.remoteEndPoint);
-                if (this.totalLength == 0)
-                {
-                    return NetworkReadResult.CreateDisconnected();
-                }
+        _buffer = new byte[bufferSize];
+    }
 
-                return NetworkReadResult.CreateSuccess((IPEndPoint)this.remoteEndPoint);
-            }
-            catch (SocketException e)
+    public NetworkReadResult BeginReading()
+    {
+        _readIndex = 0;
+        try
+        {
+            _totalLength = _udpClient.ReceiveFrom(_buffer, SocketFlags.None, ref _remoteEndPoint);
+            if (_totalLength == 0)
             {
-                return NetworkReadResult.CreateError(e.SocketErrorCode);
+                return NetworkReadResult.CreateDisconnected();
             }
+
+            return NetworkReadResult.CreateSuccess((IPEndPoint)_remoteEndPoint);
+        }
+        catch (SocketException e)
+        {
+            return NetworkReadResult.CreateError(e.SocketErrorCode);
+        }
+    }
+
+    public void EndReading()
+    {
+        if (_readIndex != _totalLength)
+        {
+            throw new RpcException("Not all UDP data was read.");
+        }
+    }
+
+    public ReadOnlySpan<byte> Read(int length)
+    {
+        if ((_readIndex + length) > _totalLength)
+        {
+            throw new RpcException("UDP buffer underflow.");
         }
 
-        public void EndReading()
-        {
-            if (this.readIndex != this.totalLength)
-            {
-                throw new RpcException("Not all UDP data was read.");
-            }
-        }
-
-        public ReadOnlySpan<byte> Read(int length)
-        {
-            if ((this.readIndex + length) > this.totalLength)
-            {
-                throw new RpcException("UDP buffer underflow.");
-            }
-
-            Span<byte> span = this.buffer.AsSpan(this.readIndex, length);
-            this.readIndex += length;
-            return span;
-        }
+        Span<byte> span = _buffer.AsSpan(_readIndex, length);
+        _readIndex += length;
+        return span;
     }
 }
