@@ -1,80 +1,79 @@
-namespace RpcNet
+// Copyright by Artur Wolf
+
+namespace RpcNet;
+
+using System.Net;
+using Internal;
+
+public abstract class ServerStub : IDisposable
 {
-    using System;
-    using System.Net;
-    using RpcNet.Internal;
+    protected readonly XdrVoid Void = new();
+    protected readonly ServerSettings Settings;
 
-    public abstract class ServerStub : IDisposable
+    private readonly RpcTcpServer _rpcTcpServer;
+    private readonly RpcUdpServer _rpcUdpServer;
+
+    private bool _isDisposed;
+
+    protected ServerStub(
+        Protocol protocol,
+        IPAddress ipAddress,
+        int port,
+        int program,
+        int[] versions,
+        ServerSettings serverSettings = default)
     {
-        private readonly RpcTcpServer rpcTcpServer;
-        private readonly RpcUdpServer rpcUdpServer;
-
-        private bool isDisposed;
-
-        protected ServerStub(
-            Protocol protocol,
-            IPAddress ipAddress,
-            int program,
-            int[] versions,
-            ServerSettings serverSettings = default)
+        if (ipAddress == null)
         {
-            if (ipAddress == null)
-            {
-                throw new ArgumentNullException(nameof(ipAddress));
-            }
-
-            if ((versions == null) || (versions.Length == 0))
-            {
-                throw new ArgumentNullException(nameof(versions));
-            }
-
-            if (protocol.HasFlag(Protocol.Tcp))
-            {
-                this.rpcTcpServer = new RpcTcpServer(
-                    ipAddress,
-                    program,
-                    versions,
-                    this.DispatchReceivedCall,
-                    serverSettings);
-            }
-
-            if (protocol.HasFlag(Protocol.Udp))
-            {
-                this.rpcUdpServer = new RpcUdpServer(
-                    ipAddress,
-                    program,
-                    versions,
-                    this.DispatchReceivedCall,
-                    serverSettings);
-            }
+            throw new ArgumentNullException(nameof(ipAddress));
         }
 
-        public void Start()
+        if ((versions == null) || (versions.Length == 0))
         {
-            this.rpcTcpServer?.Start();
-            this.rpcUdpServer?.Start();
+            throw new ArgumentNullException(nameof(versions));
         }
 
-        public void Dispose()
+        Settings = serverSettings;
+
+        if (protocol.HasFlag(Protocol.Tcp))
         {
-            this.Dispose(true);
-            GC.SuppressFinalize(this);
+            _rpcTcpServer = new RpcTcpServer(ipAddress, port, program, versions, DispatchReceivedCall, serverSettings);
         }
 
-        protected abstract void DispatchReceivedCall(ReceivedRpcCall call);
-
-        protected virtual void Dispose(bool disposing)
+        if (protocol.HasFlag(Protocol.Udp))
         {
-            if (!this.isDisposed)
+            _rpcUdpServer = new RpcUdpServer(ipAddress, port, program, versions, DispatchReceivedCall, serverSettings);
+        }
+    }
+
+    public int TcpPort { get; private set; }
+    public int UdpPort { get; private set; }
+
+    public void Start()
+    {
+        TcpPort = _rpcTcpServer?.Start() ?? 0;
+        UdpPort = _rpcUdpServer?.Start() ?? 0;
+    }
+
+    public void Dispose()
+    {
+        Dispose(true);
+        GC.SuppressFinalize(this);
+    }
+
+    protected abstract void DispatchReceivedCall(ReceivedRpcCall call);
+
+    protected virtual void Dispose(bool disposing)
+    {
+        if (!_isDisposed)
+        {
+            if (disposing)
             {
-                if (disposing)
-                {
-                    this.rpcUdpServer?.Dispose();
-                    this.rpcTcpServer?.Dispose();
-                }
-
-                this.isDisposed = true;
+                _rpcUdpServer?.Dispose();
+                _rpcTcpServer?.Dispose();
             }
+
+            _isDisposed = true;
         }
     }
 }
