@@ -1,103 +1,42 @@
-namespace Test
+// Copyright by Artur Wolf
+
+namespace Test;
+
+using System.Net;
+using NUnit.Framework;
+using RpcNet;
+using RpcNet.Internal;
+
+[TestFixture]
+internal class TestTcpClientServer
 {
-    using System;
-    using System.Net;
-    using NUnit.Framework;
-    using RpcNet;
-    using RpcNet.Internal;
-    using TestService;
+    private readonly IPAddress _ipAddress = IPAddress.Loopback;
 
-    [TestFixture]
-    internal class TestTcpClientServer
+    [Test]
+    public void ServerIsNotRunning()
     {
-        private const int Port = 12345;
+        const int Program = 12;
+        const int Version = 13;
 
-        private readonly IPAddress ipAddress = IPAddress.Loopback;
+        var clientSettings = new ClientSettings { Logger = new TestLogger("TCP Client") };
 
-        [Test]
-        public void ServerIsNotRunning()
-        {
-            const int Program = 12;
-            const int Version = 13;
+        RpcException exception = Assert.Throws<RpcException>(
+            () => _ = new RpcTcpClient(_ipAddress, 1, Program, Version, clientSettings));
 
-            var clientSettings = new ClientSettings
-            {
-                Port = Port,
-                Logger = new TestLogger("TCP Client")
-            };
+        Assert.That(
+            exception.Message,
+            Is.EqualTo($"Could not connect to {_ipAddress}:{1}. Socket error: ConnectionRefused."));
+    }
 
-            RpcException exception = Assert.Throws<RpcException>(
-                () => _ = new RpcTcpClient(this.ipAddress, Program, Version, clientSettings));
+    [Test]
+    public void ServerShutdownWithoutException()
+    {
+        const int Program = 12;
+        const int Version = 13;
 
-            Assert.That(
-                exception.Message,
-                Is.EqualTo($"Could not connect to {this.ipAddress}:{Port}. Socket error: ConnectionRefused."));
-        }
+        var serverSettings = new ServerSettings { Logger = new TestLogger("TCP Server") };
 
-        [Test]
-        public void ServerShutdownWithoutException()
-        {
-            const int Program = 12;
-            const int Version = 13;
-
-            var serverSettings = new ServerSettings
-            {
-                Logger = new TestLogger("TCP Server"),
-                Port = Port
-            };
-
-            var server = new RpcTcpServer(this.ipAddress, Program, new[] { Version }, call => { }, serverSettings);
-            Assert.DoesNotThrow(() => server.Dispose());
-        }
-
-        [Test]
-        public void TcpConnection()
-        {
-            const int Program = 12;
-            const int Version = 13;
-            const int Procedure = 14;
-
-            var receivedCallChannel = new Channel<ReceivedRpcCall>();
-
-            void Dispatcher(ReceivedRpcCall call)
-            {
-                // To assert it on the main thread
-                receivedCallChannel.Send(call);
-
-                var pingStruct = new PingStruct();
-                call.RetrieveCall(pingStruct);
-                call.Reply(pingStruct);
-            }
-
-            var serverSettings = new ServerSettings
-            {
-                Logger = new TestLogger("TCP Server"),
-                Port = Port
-            };
-
-            using var server = new RpcTcpServer(this.ipAddress, Program, new[] { Version }, Dispatcher, serverSettings);
-            server.Start();
-
-            for (int i = 0; i < 10; i++)
-            {
-                var clientSettings = new ClientSettings
-                {
-                    Port = Port,
-                    Logger = new TestLogger("TCP Client")
-                };
-
-                using var client = new RpcTcpClient(this.ipAddress, Program, Version, clientSettings);
-                var argument = new PingStruct { Value = i };
-                var result = new PingStruct();
-
-                client.Call(Procedure, Version, argument, result);
-
-                Assert.That(receivedCallChannel.TryReceive(TimeSpan.FromSeconds(10), out ReceivedRpcCall receivedCall));
-                Assert.That(receivedCall.Procedure, Is.EqualTo(Procedure));
-                Assert.That(receivedCall.Version, Is.EqualTo(Version));
-
-                Assert.That(argument.Value, Is.EqualTo(result.Value));
-            }
-        }
+        var server = new RpcTcpServer(_ipAddress, 0, Program, new[] { Version }, call => { }, serverSettings);
+        Assert.DoesNotThrow(() => server.Dispose());
     }
 }
