@@ -1,64 +1,58 @@
-namespace Test
+// Copyright by Artur Wolf
+
+namespace Test;
+
+using System.Net;
+using NUnit.Framework;
+using RpcNet;
+using RpcNet.Internal;
+using TestService;
+
+[TestFixture]
+internal class TestUdpClientServer : TestBase
 {
-    using System;
-    using System.Net;
-    using NUnit.Framework;
-    using RpcNet;
-    using RpcNet.Internal;
-    using TestService;
-
-    [TestFixture]
-    internal class TestUdpClientServer
+    [Test]
+    public void SendAndReceiveData()
     {
-        private const int Port = 12345;
+        IPAddress ipAddress = IPAddress.Loopback;
+        const int Program = 12;
+        const int Version = 13;
+        const int Procedure = 14;
 
-        [Test]
-        public void SendAndReceiveData()
+        var receivedCallChannel = new Channel<ReceivedRpcCall>();
+
+        void Dispatcher(ReceivedRpcCall call)
         {
-            IPAddress ipAddress = IPAddress.Loopback;
-            const int Program = 12;
-            const int Version = 13;
-            const int Procedure = 14;
+            // To assert it on the main thread
+            receivedCallChannel.Send(call);
 
-            var receivedCallChannel = new Channel<ReceivedRpcCall>();
-
-            void Dispatcher(ReceivedRpcCall call)
-            {
-                // To assert it on the main thread
-                receivedCallChannel.Send(call);
-
-                var pingStruct = new PingStruct();
-                call.RetrieveCall(pingStruct);
-                call.Reply(pingStruct);
-            }
-
-            var serverSettings = new ServerSettings
-            {
-                Logger = new TestLogger("UDP Server"),
-                Port = Port
-            };
-
-            using var server = new RpcUdpServer(ipAddress, Program, new[] { Version }, Dispatcher, serverSettings);
-            server.Start();
-
-            var clientSettings = new ClientSettings
-            {
-                Port = Port,
-                Logger = new TestLogger("UDP Client")
-            };
-
-            using var client = new RpcUdpClient(ipAddress, Program, Version, clientSettings);
-            var argument = new PingStruct { Value = 42 };
-            var result = new PingStruct();
-
-            client.Call(Procedure, Version, argument, result);
-
-            Assert.That(receivedCallChannel.TryReceive(TimeSpan.FromSeconds(10), out ReceivedRpcCall receivedCall));
-            Assert.That(receivedCall.Procedure, Is.EqualTo(Procedure));
-            Assert.That(receivedCall.Version, Is.EqualTo(Version));
-            Assert.That(receivedCall.Caller, Is.Not.Null);
-
-            Assert.That(argument.Value, Is.EqualTo(result.Value));
+            var pingStruct = new SimpleStruct();
+            call.RetrieveCall(pingStruct);
+            call.Reply(pingStruct);
         }
+
+        var serverSettings = new ServerSettings
+        {
+            Logger = new TestLogger("UDP Server"),
+            PortMapperPort = PortMapperPort
+        };
+
+        using var server = new RpcUdpServer(ipAddress, 0, Program, new[] { Version }, Dispatcher, serverSettings);
+        int port = server.Start();
+
+        var clientSettings = new ClientSettings { Logger = new TestLogger("UDP Client") };
+
+        using var client = new RpcUdpClient(ipAddress, port, Program, Version, clientSettings);
+        var argument = new SimpleStruct { Value = 42 };
+        var result = new SimpleStruct();
+
+        client.Call(Procedure, Version, argument, result);
+
+        Assert.That(receivedCallChannel.TryReceive(TimeSpan.FromSeconds(10), out ReceivedRpcCall receivedCall));
+        Assert.That(receivedCall.Procedure, Is.EqualTo(Procedure));
+        Assert.That(receivedCall.Version, Is.EqualTo(Version));
+        Assert.That(receivedCall.Caller, Is.Not.Null);
+
+        Assert.That(argument.Value, Is.EqualTo(result.Value));
     }
 }
