@@ -1,13 +1,13 @@
-namespace RpcNet.Internal
-{
-    using System;
-    using System.Net;
+// Copyright by Artur Wolf
 
-    // Public for tests
-    public interface INetworkWriter
-    {
-        void BeginWriting();
-        NetworkWriteResult EndWriting(IPEndPoint remoteEndPoint);
-        Span<byte> Reserve(int length);
-    }
+namespace RpcNet.Internal;
+
+using System.Net;
+
+// Public for tests
+public interface INetworkWriter
+{
+    void BeginWriting();
+    NetworkWriteResult EndWriting(IPEndPoint remoteEndPoint);
+    Span<byte> Reserve(int length);
 }
