@@ -1,13 +1,12 @@
-namespace RpcNet
-{
-    using System;
+// Copyright by Artur Wolf
 
-    [Flags]
-    public enum Protocol
-    {
-        Unknown = 0,
-        Tcp = 1,
-        Udp = 2,
-        TcpAndUdp = Tcp | Udp
-    }
+namespace RpcNet;
+
+[Flags]
+public enum Protocol
+{
+    Unknown = 0,
+    Tcp = 1,
+    Udp = 2,
+    TcpAndUdp = Tcp | Udp
 }
